# swagger.yaml
openapi: 3.0.3
info:
  title: Image and Auth API
  version: 1.0.0
  description: API documentation for image upload and user authentication.
paths:
...
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
              - username
              - email
              - password
      responses:
        '201':
          description: User created
        '400':
          description: Missing data
        '409':
          description: Username or email already exists
        '500':
          description: Server error
  /login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
              - username
              - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Missing credentials
        '401':
          description: Invalid credentials
        '500':
          description: Server error
  /logout:
    post:
      summary: Log user out
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Wylogowano pomyślnie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jwt=; Max-Age=0; SameSite=Strict; Path=/, csrfToken=; Max-Age=0; SameSite=Strict; Path=/
              description: Deletes JWT and CSRF token cookies
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Wystąpił błąd serwera
  /verify:
    post:
      summary: Verify JWT token
      parameters:
      - name: Cookie
        in: header
        required: true
        schema:
          type: string
          example: jwt=<token>
        description: JWT token stored in cookie
      responses:
        '200':
          description: Token valid
        '401':
          description: Invalid or missing token
        '500':
          description: Server error
...